WIP use base network view for paint - base network view emits a 'network-loaded' once the graph has been loaded and the initial layout has finished - paint data on 'network-loaded' - todo - create paint menu component that is hooked up to the paint button, (also need to tinker with passing enrichment data to this menu)
