+++ src/client/features/paint/paint-menu.js
@@ -1,0 +1,7 @@
+const React = require('react');
+const h = require('react-hyperscript');
+
+
+class PaintMenu extends React.Component {
+  
+}
